Refactor some parts of the FlightAPI implementation
