+++ src/main/java/com/example/flight/controllers/AirportController.java
@@ -2,7 +2,7 @@ package com.example.flight.controllers;
 
 import java.util.List;
 
-import com.example.flight.Data;
+import com.example.flight.domains.Data;
 import com.example.flight.entities.Airport;
 import com.example.flight.entities.Flight;
 import com.example.flight.services.AirportService;
@@ -70,10 +70,8 @@ class AirportController {
         flightService.deleteFlight(id);
     }
 
-    //The endpoint below is for pure testing
-
-    @GetMapping("/test/{searchParam}")
+    @GetMapping("/flightAPI/{searchParam}")
     public Data getAirportsByNameTest(@PathVariable String searchParam) {
-        return airportService.getAirportsTest(searchParam);
+        return airportService.getAirportsByParam(searchParam);
     }
 }
