+++ src/main/java/com/example/flight/domains/Airport.java
@@ -1,4 +1,4 @@
-package com.example.flight;
+package com.example.flight.domains;
 
 import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
 import lombok.NoArgsConstructor;
@@ -6,6 +6,8 @@ import lombok.NoArgsConstructor;
 @lombok.Data
 @NoArgsConstructor
 @JsonIgnoreProperties(ignoreUnknown = true)
-public class Data {
-    private AirportTest[] data;
+public class Airport {
+    private String iata;
+    private String fs;
+    private String name;
 }
